[jupyter] refactor into module
