+++ src/pymor/gui/jupyter/__init__.py
@@ -1,0 +1,82 @@
+# This file is part of the pyMOR project (http://www.pymor.org).
+# Copyright 2013-2019 pyMOR developers and contributors. All rights reserved.
+# License: BSD 2-Clause License (http://opensource.org/licenses/BSD-2-Clause)
+
+"""This module provides plotting support inside the Jupyter notebook.
+
+To use these routines you first have to execute ::
+
+        %matplotlib notebook
+
+inside the given notebook.
+"""
+import IPython
+
+from pymor.gui.jupyter.patch import visualize_patch
+
+default_visualizer = visualize_patch
+
+
+def progress_bar(sequence, every=None, size=None, name='Parameters'):
+    from ipywidgets import IntProgress, HTML, VBox
+    # c&p from https://github.com/kuk/log-progress
+    is_iterator = False
+    if size is None:
+        try:
+            size = len(sequence)
+        except TypeError:
+            is_iterator = True
+    if size is not None:
+        if every is None:
+            if size <= 200:
+                every = 1
+            else:
+                every = int(size / 200)     # every 0.5%
+    else:
+        assert every is not None, 'sequence is iterator, set every'
+
+    if is_iterator:
+        progress = IntProgress(min=0, max=1, value=1)
+        progress.bar_style = 'info'
+    else:
+        progress = IntProgress(min=0, max=size, value=0)
+    label = HTML()
+    box = VBox(children=[label, progress])
+    IPython.display.display(box)
+
+    index = 0
+    try:
+        for index, record in enumerate(sequence, 1):
+            if index == 1 or index % every == 0:
+                if is_iterator:
+                    label.value = '{name}: {index} / ?'.format(
+                        name=name,
+                        index=index
+                    )
+                else:
+                    progress.value = index
+                    label.value = u'{name}: {index} / {size}'.format(
+                        name=name,
+                        index=index,
+                        size=size
+                    )
+            yield record
+    except:
+        progress.bar_style = 'danger'
+        raise
+    else:
+        progress.bar_style = 'success'
+        progress.value = index
+        label.value = f"{name}: {str(index or '?')}"
+
+
+def load_ipython_extension(ipython):
+    from pymor.gui.jupyter.logging import redirect_logging
+    ipython.events.register('pre_run_cell', redirect_logging.start)
+    ipython.events.register('post_run_cell', redirect_logging.stop)
+
+
+def unload_ipython_extension(ipython):
+    from pymor.gui.jupyter.logging import redirect_logging
+    ipython.events.unregister('pre_run_cell', redirect_logging.start)
+    ipython.events.unregister('post_run_cell', redirect_logging.stop)
